% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw_07_rows.R
\name{rm_0s_by_row}
\alias{rm_0s_by_row}
\title{Homework 07
Removing rows from a data set}
\usage{
rm_0s_by_row(data, ...)
}
\arguments{
\item{data}{Input, must be a matrix.}

\item{...}{Additional arguments to be passed to `[sub]` function.}
}
\value{
The new matrix without the rows with only 0's
}
\description{
Here we have a function that will remove rows with all 0s in a matrix.
}
\examples{
# Data set:

set.seed(123)
x = matrix(rpois(100, 0.1), 10, 10)

}
